+++ sip-core/src/main/java/de/ikor/sip/foundation/core/framework/connectors/OutConnector.java
@@ -5,7 +5,7 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.model.OnExceptionDefinition;
 import org.apache.camel.model.RouteDefinition;
 
-public abstract class OutConnectorDefinition extends Connector {
+public abstract class OutConnector extends Connector {
 
   @Setter private RouteBuilder routeBuilder;
