+++ sip-core/src/test/java/de/ikor/sip/foundation/core/apps/framework/ConfigurationTestingCentralRouter.java
@@ -2,13 +2,13 @@ package de.ikor.sip.foundation.core.apps.framework;
 
 import static org.apache.camel.builder.Builder.simple;
 
-import de.ikor.sip.foundation.core.framework.routers.CentralRouterDefinition;
+import de.ikor.sip.foundation.core.framework.routers.CentralRouter;
 import de.ikor.sip.foundation.core.framework.routers.IntegrationScenario;
 import de.ikor.sip.foundation.core.framework.stubs.ConfigInConnector;
 import de.ikor.sip.foundation.core.framework.stubs.ConfigOutConnector;
 
 @IntegrationScenario(name = "config-scenario")
-public class ConfigurationTestingCentralRouter extends CentralRouterDefinition {
+public class ConfigurationTestingCentralRouter extends CentralRouter {
 
   public static final String SCENARIO_HEADER_KEY = "scenario";
   public static final String SCENARIO_HEADER_VALUE = "scenario header";
