+++ sip-core/src/test/java/de/ikor/sip/foundation/core/apps/framework/centralrouter/CentralRouterTestingApplication.java
@@ -1,16 +1,16 @@
 package de.ikor.sip.foundation.core.apps.framework.centralrouter;
 
 import de.ikor.sip.foundation.core.annotation.SIPIntegrationAdapter;
-import de.ikor.sip.foundation.core.framework.routers.CentralRouterDefinition;
-import de.ikor.sip.foundation.core.framework.stubs.TestingCentralRouterDefinition;
+import de.ikor.sip.foundation.core.framework.routers.CentralRouter;
+import de.ikor.sip.foundation.core.framework.stubs.TestingCentralRouter;
 import org.apache.camel.builder.RouteBuilder;
 import org.springframework.context.annotation.Bean;
 
 @SIPIntegrationAdapter
 public class CentralRouterTestingApplication {
   @Bean
-  CentralRouterDefinition testingCentralRouter() {
-    return new TestingCentralRouterDefinition();
+  CentralRouter testingCentralRouter() {
+    return new TestingCentralRouter();
   }
 
   @Bean
