+++ sip-core/src/test/java/de/ikor/sip/foundation/core/apps/framework/restrouter/ScopeBeanCentralRouter.java
@@ -1,6 +1,6 @@
 package de.ikor.sip.foundation.core.apps.framework.restrouter;
 
-import de.ikor.sip.foundation.core.framework.routers.CentralRouterDefinition;
+import de.ikor.sip.foundation.core.framework.routers.CentralRouter;
 import de.ikor.sip.foundation.core.framework.routers.IntegrationScenario;
 import de.ikor.sip.foundation.core.framework.stubs.ScopeAppendOutConnector;
 import de.ikor.sip.foundation.core.framework.stubs.ScopeBeanInConnector;
@@ -8,7 +8,7 @@ import org.apache.camel.converter.stream.InputStreamCache;
 import org.springframework.beans.factory.annotation.Autowired;
 
 @IntegrationScenario(name = "bean-scenario", responseType = String.class, requestType = InputStreamCache.class)
-public class ScopeBeanCentralRouterDefinition extends CentralRouterDefinition {
+public class ScopeBeanCentralRouter extends CentralRouter {
   @Autowired
   private ScopeAppendOutConnector appendStringOutConnector;
   @Autowired
