+++ sip-core/src/test/java/de/ikor/sip/foundation/core/framework/RestConnectorIntegrationTest.java
@@ -3,7 +3,7 @@ package de.ikor.sip.foundation.core.framework;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import de.ikor.sip.foundation.core.apps.framework.centralrouter.CentralRouterTestingApplication;
-import de.ikor.sip.foundation.core.apps.framework.restrouter.RestCentralRouterDefinition;
+import de.ikor.sip.foundation.core.apps.framework.restrouter.RestCentralRouter;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -13,7 +13,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(
-    classes = {CentralRouterTestingApplication.class, RestCentralRouterDefinition.class},
+    classes = {CentralRouterTestingApplication.class, RestCentralRouter.class},
     webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @DirtiesContext
 class RestConnectorIntegrationTest {
