+++ sip-core/src/test/java/de/ikor/sip/foundation/core/framework/stubs/AppendStringOutConnector.java
@@ -1,9 +1,9 @@
 package de.ikor.sip.foundation.core.framework.stubs;
 
-import de.ikor.sip.foundation.core.framework.connectors.OutConnectorDefinition;
+import de.ikor.sip.foundation.core.framework.connectors.OutConnector;
 import org.apache.camel.model.RouteDefinition;
 
-public class AppendStringOutConnector extends OutConnectorDefinition {
+public class AppendStringOutConnector extends OutConnector {
   @Override
   public void configure(RouteDefinition route) {
     route.process(
