+++ sip-core/src/test/java/de/ikor/sip/foundation/core/framework/stubs/ComplexOutConnector.java
@@ -1,6 +1,6 @@
 package de.ikor.sip.foundation.core.framework.stubs;
 
-import de.ikor.sip.foundation.core.framework.connectors.OutConnectorDefinition;
+import de.ikor.sip.foundation.core.framework.connectors.OutConnector;
 import de.ikor.sip.foundation.core.framework.endpoints.OutEndpoint;
 import org.apache.camel.AggregationStrategy;
 import org.apache.camel.Message;
@@ -8,7 +8,7 @@ import org.apache.camel.RoutesBuilder;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.model.RouteDefinition;
 
-public class ComplexOutConnector extends OutConnectorDefinition {
+public class ComplexOutConnector extends OutConnector {
 
   public static final String DIRECT_COMPLEX_MCAST_1_URI = "direct:complex-mcast-1";
   public static final String DIRECT_COMPLEX_MCAST_2_URI = "direct:complex-mcast-2";
