+++ sip-core/src/test/java/de/ikor/sip/foundation/core/framework/stubs/ConfigInConnector.java
@@ -1,11 +1,11 @@
 package de.ikor.sip.foundation.core.framework.stubs;
 
-import de.ikor.sip.foundation.core.framework.connectors.InConnectorDefinition;
+import de.ikor.sip.foundation.core.framework.connectors.InConnector;
 import de.ikor.sip.foundation.core.framework.endpoints.InEndpoint;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
-public class ConfigInConnector extends InConnectorDefinition {
+public class ConfigInConnector extends InConnector {
   private String channel = "config";
 
   @Override
