+++ sip-core/src/test/java/de/ikor/sip/foundation/core/framework/stubs/EndpointDSLInConnector.java
@@ -1,9 +1,9 @@
 package de.ikor.sip.foundation.core.framework.stubs;
 
-import de.ikor.sip.foundation.core.framework.connectors.InConnectorDefinition;
+import de.ikor.sip.foundation.core.framework.connectors.InConnector;
 import de.ikor.sip.foundation.core.framework.endpoints.InEndpoint;
 
-public class EndpointDSLInConnector extends InConnectorDefinition {
+public class EndpointDSLInConnector extends InConnector {
 
     @Override
     public String getName() {
