+++ sip-core/src/test/java/de/ikor/sip/foundation/core/framework/stubs/InConnectorStub.java
@@ -1,6 +1,6 @@
 package de.ikor.sip.foundation.core.framework.stubs;
 
-import de.ikor.sip.foundation.core.framework.connectors.InConnectorDefinition;
+import de.ikor.sip.foundation.core.framework.connectors.InConnector;
 import de.ikor.sip.foundation.core.framework.endpoints.InEndpoint;
 import lombok.AllArgsConstructor;
 
@@ -8,7 +8,7 @@ import static org.apache.camel.builder.Builder.body;
 import static org.apache.camel.builder.Builder.simple;
 
 @AllArgsConstructor
-public class InConnectorStub extends InConnectorDefinition {
+public class InConnectorStub extends InConnector {
 
   private InEndpoint inEndpoint;
