+++ sip-core/src/test/java/de/ikor/sip/foundation/core/framework/stubs/OutConnectorStub.java
@@ -1,12 +1,12 @@
 package de.ikor.sip.foundation.core.framework.stubs;
 
-import de.ikor.sip.foundation.core.framework.connectors.OutConnectorDefinition;
+import de.ikor.sip.foundation.core.framework.connectors.OutConnector;
 import de.ikor.sip.foundation.core.framework.endpoints.OutEndpoint;
 import lombok.AllArgsConstructor;
 import org.apache.camel.model.RouteDefinition;
 
 @AllArgsConstructor
-public class OutConnectorStub extends OutConnectorDefinition {
+public class OutConnectorStub extends OutConnector {
 
     private OutEndpoint outEndpoint;
