+++ sip-core/src/test/java/de/ikor/sip/foundation/core/framework/stubs/ScopeAppendOutConnector.java
@@ -1,13 +1,13 @@
 package de.ikor.sip.foundation.core.framework.stubs;
 
 import de.ikor.sip.foundation.core.framework.beans.CDMHolder;
-import de.ikor.sip.foundation.core.framework.connectors.OutConnectorDefinition;
+import de.ikor.sip.foundation.core.framework.connectors.OutConnector;
 import org.apache.camel.model.RouteDefinition;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ScopeAppendOutConnector extends OutConnectorDefinition {
+public class ScopeAppendOutConnector extends OutConnector {
     @Autowired
     CDMHolder bean;
